+++ src/backend/context.rs
@@ -2,7 +2,7 @@ use core::num::{NonZeroU32, NonZeroU64};
 
 use crate::backend::proxy::ContextFunction;
 use crate::backend::util::GpuImage;
-use crate::backend::util::UniformsBuffer;
+use crate::backend::util::{ToStd140, UniformsBuffer};
 use crate::backend::window::Window;
 use crate::backend::window::WindowUniforms;
 use crate::background_thread::BackgroundThread;
@@ -10,15 +10,13 @@ use crate::error::CreateWindowError;
 use crate::error::GetDeviceError;
 use crate::error::InvalidWindowId;
 use crate::error::NoSuitableAdapterFound;
-use crate::error::SetImageError;
 use crate::event::{self, Event, EventHandlerControlFlow, WindowEvent};
-use crate::AsImageView;
 use crate::ContextProxy;
-use crate::ImageInfo;
-use crate::Rectangle;
+use crate::ImageView;
 use crate::WindowHandle;
 use crate::WindowId;
 use crate::WindowOptions;
+use glam::Affine2;
 
 /// Internal shorthand type-alias for the correct [`winit::event_loop::EventLoop`].
 ///
@@ -26,6 +24,8 @@ use crate::WindowOptions;
 type EventLoop = winit::event_loop::EventLoop<ContextFunction>;
 
 /// Internal shorthand for context event handlers.
+///
+/// Not for use in public APIs.
 type DynContextEventHandler = dyn FnMut(&mut ContextHandle, &mut Event, &mut event::EventHandlerControlFlow);
 
 /// Internal shorthand type-alias for the correct [`winit::event_loop::EventLoopWindowTarget`].
@@ -103,8 +103,8 @@ pub struct Context {
 /// You can interact with the global context through a [`ContextHandle`] only from the global context thread.
 /// To interact with the context from a different thread, use a [`ContextProxy`].
 pub struct ContextHandle<'a> {
-	context: &'a mut Context,
-	event_loop: &'a EventLoopWindowTarget,
+	pub(crate) context: &'a mut Context,
+	pub(crate) event_loop: &'a EventLoopWindowTarget,
 }
 
 impl Context {
@@ -119,6 +119,9 @@ impl Context {
 		let proxy = ContextProxy::new(event_loop.create_proxy(), std::thread::current().id());
 
 		let (device, queue) = futures::executor::block_on(get_device(&instance))?;
+		device.on_uncaptured_error(|error| {
+			panic!("Unhandled WGPU error: {}", error);
+		});
 
 		let window_bind_group_layout = create_window_bind_group_layout(&device);
 		let image_bind_group_layout = create_image_bind_group_layout(&device);
@@ -175,21 +178,6 @@ impl Context {
 		self.event_handlers.push(Box::new(handler))
 	}
 
-	/// Add a window-specific event handler.
-	pub fn add_window_event_handler<F>(&mut self, window_id: WindowId, handler: F) -> Result<(), InvalidWindowId>
-	where
-		F: 'static + FnMut(&mut WindowHandle, &mut WindowEvent, &mut EventHandlerControlFlow),
-	{
-		let window = self
-			.windows
-			.iter_mut()
-			.find(|x| x.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-
-		window.event_handlers.push(Box::new(handler));
-		Ok(())
-	}
-
 	/// Run the event loop of the context.
 	///
 	/// This function must be run from the main thread and never returns.
@@ -242,120 +230,16 @@ impl<'a> ContextHandle<'a> {
 		self.context.exit_with_last_window = exit_with_last_window;
 	}
 
+	/// Get a window handle for the given window ID.
+	pub fn window(&mut self, window_id: WindowId) -> Result<WindowHandle, InvalidWindowId> {
+		let index = self.context.windows.iter().position(|x| x.id() == window_id).ok_or(InvalidWindowId { window_id })?;
+		Ok(WindowHandle::new(self.reborrow(), index))
+	}
+
 	/// Create a new window.
 	pub fn create_window(&mut self, title: impl Into<String>, options: WindowOptions) -> Result<WindowHandle, CreateWindowError> {
-		let window_id = self.context.create_window(self.event_loop, title, options)?;
-		Ok(WindowHandle::new(self.reborrow(), window_id))
-	}
-
-	/// Destroy a window.
-	pub fn destroy_window(&mut self, window_id: WindowId) -> Result<(), InvalidWindowId> {
-		self.context.destroy_window(window_id)
-	}
-
-	/// Get the image info and the area where the image is drawn for a window.
-	pub fn window_image_info(&self, window_id: WindowId) -> Result<Option<(ImageInfo, Rectangle)>, InvalidWindowId> {
-		let window = self.context.windows.iter().find(|x| x.id() == window_id).ok_or(InvalidWindowId { window_id })?;
-		let image_info = match window.image.as_ref().map(|x| *x.info()) {
-			Some(x) => x,
-			None => return Ok(None),
-		};
-
-		let uniforms = window.calculate_uniforms();
-		let window_size = window.window.inner_size();
-
-		let [x, y] = uniforms.offset;
-		let [width, height] = uniforms.relative_size;
-
-		let x = (x * window_size.width as f32) as i32;
-		let y = (y * window_size.height as f32) as i32;
-		let width = (width * window_size.width as f32) as u32;
-		let height = (height * window_size.height as f32) as u32;
-
-		let image_area = Rectangle::from_xywh(x, y, width, height);
-
-		Ok(Some((image_info, image_area)))
-	}
-
-	/// Make a window visible or invisible.
-	pub fn set_window_visible(&mut self, window_id: WindowId, visible: bool) -> Result<(), InvalidWindowId> {
-		self.context.set_window_visible(window_id, visible)
-	}
-
-	/// Change the options of a window.
-	pub fn set_window_options<F>(&mut self, window_id: WindowId, make_options: F) -> Result<(), InvalidWindowId>
-	where
-		F: FnOnce(&WindowOptions) -> WindowOptions,
-	{
-		let window = self
-			.context
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-		let options = (make_options)(&window.options);
-
-		window.window.set_resizable(options.resizable);
-		window.window.set_decorations(!options.borderless);
-		if options.size != window.options.size {
-			if let Some(size) = options.size {
-				window.window.set_inner_size(winit::dpi::LogicalSize::<u32>::from(size));
-			}
-		}
-
-		window.options = options;
-		window.window.request_redraw();
-		Ok(())
-	}
-
-	/// Set the image to be displayed on a window.
-	pub fn set_window_image(
-		&mut self,
-		window_id: WindowId,
-		name: impl Into<String>,
-		image: &impl AsImageView,
-	) -> Result<(), SetImageError> {
-		self.context.set_window_image(window_id, name.into(), image)
-	}
-
-	/// Add an overlay to a window.
-	///
-	/// Overlays are drawn on top of the image.
-	/// Overlays remain active until you call they are cleared.
-	pub fn add_window_overlay(
-		&mut self,
-		window_id: WindowId,
-		name: impl Into<String>,
-		image: &impl AsImageView,
-	) -> Result<(), SetImageError> {
-		let window = self
-			.context
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-		let image = GpuImage::from_data(
-			name.into(),
-			&self.context.device,
-			&self.context.image_bind_group_layout,
-			image.as_image_view()?,
-		);
-		window.overlays.push(image);
-		window.window.request_redraw();
-		Ok(())
-	}
-
-	/// Clear the overlays of a window.
-	pub fn clear_window_overlays(&mut self, window_id: WindowId) -> Result<(), InvalidWindowId> {
-		let window = self
-			.context
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-		window.overlays.clear();
-		window.window.request_redraw();
-		Ok(())
+		let index = self.context.create_window(self.event_loop, title, options)?;
+		Ok(WindowHandle::new(self.reborrow(), index))
 	}
 
 	/// Add a global event handler.
@@ -364,14 +248,6 @@ impl<'a> ContextHandle<'a> {
 		F: 'static + FnMut(&mut ContextHandle, &mut Event, &mut EventHandlerControlFlow),
 	{
 		self.context.add_event_handler(handler);
-	}
-
-	/// Add a window-specific event handler.
-	pub fn add_window_event_handler<F>(&mut self, window_id: WindowId, handler: F) -> Result<(), InvalidWindowId>
-	where
-		F: 'static + FnMut(&mut WindowHandle, &mut WindowEvent, &mut EventHandlerControlFlow),
-	{
-		self.context.add_window_event_handler(window_id, handler)
 	}
 
 	/// Run a task in a background thread and register it with the context.
@@ -406,7 +282,7 @@ impl Context {
 		event_loop: &EventLoopWindowTarget,
 		title: impl Into<String>,
 		options: WindowOptions,
-	) -> Result<WindowId, CreateWindowError> {
+	) -> Result<usize, CreateWindowError> {
 		let mut window = winit::window::WindowBuilder::new()
 			.with_title(title)
 			.with_visible(!options.start_hidden)
@@ -414,32 +290,36 @@ impl Context {
 			.with_decorations(!options.borderless);
 
 		if let Some(size) = options.size {
-			let size = winit::dpi::LogicalSize::new(size[0], size[1]);
-			window = window.with_inner_size(size);
+			window = window.with_inner_size(winit::dpi::PhysicalSize::new(size[0], size[1]));
 		}
 
 		let window = window.build(event_loop)?;
 
 		let surface = unsafe { self.instance.create_surface(&window) };
-		let swap_chain = create_swap_chain(window.inner_size(), &surface, self.swap_chain_format, &self.device);
+		let size = glam::UVec2::new(window.inner_size().width, window.inner_size().height);
+		let swap_chain = create_swap_chain(size, &surface, self.swap_chain_format, &self.device);
 		let uniforms = UniformsBuffer::from_value(&self.device, &WindowUniforms::no_image(), &self.window_bind_group_layout);
 
 		let window = Window {
 			window,
-			options,
+			preserve_aspect_ratio: options.preserve_aspect_ratio,
+			background_color: options.background_color,
 			surface,
 			swap_chain,
 			uniforms,
 			image: None,
-			zoom: 1.0,
-			translate: [0.0, 0.0],
+			user_transform: Affine2::IDENTITY,
 			overlays: Vec::new(),
+			overlays_visible: options.overlays_visible,
 			event_handlers: Vec::new(),
 		};
 
-		let window_id = window.id();
 		self.windows.push(window);
-		Ok(window_id)
+		let index = self.windows.len() - 1;
+		if options.default_controls {
+			self.windows[index].event_handlers.push(Box::new(super::window::default_controls_handler));
+		}
+		Ok(index)
 	}
 
 	/// Destroy a window.
@@ -453,34 +333,13 @@ impl Context {
 		Ok(())
 	}
 
-	/// Make a window visible or invisible.
-	fn set_window_visible(&mut self, window_id: WindowId, visible: bool) -> Result<(), InvalidWindowId> {
-		let window = self
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-		window.set_visible(visible);
-		Ok(())
-	}
-
-	/// Set the image to be displayed on a window.
-	fn set_window_image(&mut self, window_id: WindowId, name: String, image: &impl AsImageView) -> Result<(), SetImageError> {
-		let window = self
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-
-		let image = GpuImage::from_data(name, &self.device, &self.image_bind_group_layout, image.as_image_view()?);
-		window.image = Some(image);
-		window.uniforms.mark_dirty(true);
-		window.window.request_redraw();
-		Ok(())
+	/// Upload an image to the GPU.
+	pub fn make_gpu_image(&self, name: impl Into<String>, image: &ImageView) -> GpuImage {
+		GpuImage::from_data(name.into(), &self.device, &self.image_bind_group_layout, image)
 	}
 
 	/// Resize a window.
-	fn resize_window(&mut self, window_id: WindowId, new_size: winit::dpi::PhysicalSize<u32>) -> Result<(), InvalidWindowId> {
+	fn resize_window(&mut self, window_id: WindowId, new_size: glam::UVec2) -> Result<(), InvalidWindowId> {
 		let window = self
 			.windows
 			.iter_mut()
@@ -489,53 +348,6 @@ impl Context {
 
 		window.swap_chain = create_swap_chain(new_size, &window.surface, self.swap_chain_format, &self.device);
 		window.uniforms.mark_dirty(true);
-		Ok(())
-	}
-
-	/// Zoom a window.
-	fn zoom_window(
-		&mut self,
-		window_id: WindowId,
-		delta: f32,
-		mouse_position_x: f32,
-		mouse_position_y: f32
-	) -> Result<(), InvalidWindowId> {
-		let window = self
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-
-		let uniforms = window.calculate_uniforms();
-		let size = window.window.inner_size();
-		let zoom_factor = if delta > 0.0 { 1.1 } else { 1.0 / 1.1 };
-		window.translate[0] += ((mouse_position_x / size.width as f32) - uniforms.offset[0]) * (1.0 - zoom_factor);
-		window.translate[1] += (1.0 - (mouse_position_y / size.height as f32) - uniforms.offset[1]) * (1.0 - zoom_factor);
-		window.zoom *= zoom_factor;
-		window.uniforms.mark_dirty(true);
-		window.window.request_redraw();
-		Ok(())
-	}
-
-	/// Pan a window.
-	fn pan_window(
-		&mut self,
-		window_id: WindowId,
-		delta_position_x: f32,
-		delta_position_y: f32,
-	) -> Result<(), InvalidWindowId> {
-		let window = self
-			.windows
-			.iter_mut()
-			.find(|w| w.id() == window_id)
-			.ok_or(InvalidWindowId { window_id })?;
-
-		let size = window.window.inner_size();
-		window.translate[0] += delta_position_x / size.width as f32;
-		// positive image y-axis is equivalent to negative y-axis of the mouse cursor, hence subtract.
-		window.translate[1] -= delta_position_y / size.height as f32;
-		window.uniforms.mark_dirty(true);
-		window.window.request_redraw();
 		Ok(())
 	}
 
@@ -570,10 +382,10 @@ impl Context {
 			&self.window_pipeline,
 			&window.uniforms,
 			image,
-			Some(window.options.background_color),
+			Some(window.background_color),
 			&frame.output.view,
 		);
-		if window.options.show_overlays {
+		if window.overlays_visible {
 			for overlay in &window.overlays {
 				render_pass(
 					&mut encoder,
@@ -601,18 +413,17 @@ impl Context {
 			None => return Ok(None),
 		};
 
-		let bytes_per_row = align_next_u32(image.info().width * 4, wgpu::COPY_BYTES_PER_ROW_ALIGNMENT);
+		let bytes_per_row = align_next_u32(image.info().size.x * 4, wgpu::COPY_BYTES_PER_ROW_ALIGNMENT);
 
 		let size = wgpu::Extent3d {
 			width: div_round_up(bytes_per_row, 4),
-			height: image.info().height,
+			height: image.info().size.y,
 			depth_or_array_layers: 1,
 		};
 
 		let window_uniforms = WindowUniforms {
-			offset: [0.0, 0.0],
-			relative_size: [image.info().width as f32 / size.width as f32, 1.0],
-			pixel_size: [image.info().width as f32, image.info().height as f32],
+			transform: Affine2::IDENTITY,
+			image_size: image.info().size.as_f32(),
 		};
 		let window_uniforms = UniformsBuffer::from_value(&self.device, &window_uniforms, &self.window_bind_group_layout);
 
@@ -655,7 +466,7 @@ impl Context {
 
 		let buffer = self.device.create_buffer(&wgpu::BufferDescriptor {
 			label: None,
-			size: u64::from(bytes_per_row) * u64::from(image.info().height),
+			size: u64::from(bytes_per_row) * u64::from(image.info().size.y),
 			usage: wgpu::BufferUsage::COPY_DST | wgpu::BufferUsage::MAP_READ,
 			mapped_at_creation: false,
 		});
@@ -671,7 +482,7 @@ impl Context {
 				layout: wgpu::ImageDataLayout {
 					offset: 0,
 					bytes_per_row: NonZeroU32::new(bytes_per_row),
-					rows_per_image: NonZeroU32::new(image.info().height),
+					rows_per_image: NonZeroU32::new(image.info().size.y),
 				},
 			},
 			size,
@@ -682,10 +493,8 @@ impl Context {
 		let view = super::util::map_buffer(&self.device, buffer.slice(..)).unwrap();
 		let info = crate::ImageInfo {
 			pixel_format: crate::PixelFormat::Rgba8(crate::Alpha::Unpremultiplied),
-			width: image.info().width,
-			height: image.info().height,
-			stride_x: 4,
-			stride_y: bytes_per_row,
+			size: image.info().size,
+			stride: glam::UVec2::new(4, bytes_per_row),
 		};
 		let data: Box<[u8]> = Box::from(&view[..]);
 		Ok(Some((image.name().to_string(), crate::BoxImage::new(info, data))))
@@ -749,28 +558,8 @@ impl Context {
 				}
 			},
 			Event::WindowEvent(WindowEvent::Resized(event)) => {
-				if event.size.width > 0 && event.size.height > 0 {
+				if event.size.x > 0 && event.size.y > 0 {
 					let _ = self.resize_window(event.window_id, event.size);
-				}
-			},
-			Event::WindowEvent(WindowEvent::MouseWheel(event)) => {
-				let delta = match event.delta {
-					winit::event::MouseScrollDelta::LineDelta(_, y) => y,
-					_ => 0.0
-				};
-				let current_position = self.mouse_cache.get_position(event.window_id, event.device_id).unwrap_or_else(|| [0.0, 0.0].into());
-				let _ = self.zoom_window(event.window_id, delta, current_position.x as f32, current_position.y as f32);
-			},
-			Event::WindowEvent(WindowEvent::MouseMove(event)) => {
-				if event.buttons.is_pressed(event::MouseButton::Left) {
-					let current_position = self.mouse_cache.get_position(event.window_id, event.device_id).unwrap_or_else(|| [0.0, 0.0].into());
-					let prev_position = self.mouse_cache.get_previous_position(event.window_id, event.device_id).unwrap_or_else(|| [0.0, 0.0].into());
-
-					let _ = self.pan_window(
-						event.window_id,
-						(current_position.x - prev_position.x) as f32,
-						(current_position.y - prev_position.y) as f32,
-					);
 				}
 			},
 			Event::WindowEvent(WindowEvent::RedrawRequested(event)) => {
@@ -792,7 +581,7 @@ impl Context {
 		// That's not allowed, of course, so temporarily swap the event handlers with a new vector.
 		// When we've run all handlers, we add the new handlers to the original vector and place it back.
 		// https://newfastuff.com/wp-content/uploads/2019/05/dVIkgAf.png
-		let mut event_handlers = std::mem::replace(&mut self.event_handlers, Vec::new());
+		let mut event_handlers = std::mem::take(&mut self.event_handlers);
 
 		let mut stop_propagation = false;
 		event_handlers.retain_mut(|handler| {
@@ -807,7 +596,7 @@ impl Context {
 			}
 		});
 
-		let new_event_handlers = std::mem::replace(&mut self.event_handlers, Vec::new());
+		let new_event_handlers = std::mem::take(&mut self.event_handlers);
 		event_handlers.extend(new_event_handlers);
 		self.event_handlers = event_handlers;
 	}
@@ -821,7 +610,7 @@ impl Context {
 			None => return true,
 		};
 
-		let mut event_handlers = std::mem::replace(&mut self.windows[window_index].event_handlers, Vec::new());
+		let mut event_handlers = std::mem::take(&mut self.windows[window_index].event_handlers);
 
 		let mut stop_propagation = false;
 		event_handlers.retain_mut(|handler| {
@@ -829,15 +618,15 @@ impl Context {
 				false
 			} else {
 				let context_handle = ContextHandle::new(self, event_loop);
-				let mut window_handle = WindowHandle::new(context_handle, event.window_id());
+				let window_handle = WindowHandle::new(context_handle, window_index);
 				let mut control = EventHandlerControlFlow::default();
-				(handler)(&mut window_handle, event, &mut control);
+				(handler)(window_handle, event, &mut control);
 				stop_propagation = control.stop_propagation;
 				!control.remove_handler
 			}
 		});
 
-		let new_event_handlers = std::mem::replace(&mut self.windows[window_index].event_handlers, Vec::new());
+		let new_event_handlers = std::mem::take(&mut self.windows[window_index].event_handlers);
 		event_handlers.extend(new_event_handlers);
 		self.windows[window_index].event_handlers = event_handlers;
 
@@ -859,7 +648,7 @@ impl Context {
 
 	/// Join all background tasks.
 	fn join_background_tasks(&mut self) {
-		for task in std::mem::replace(&mut self.background_tasks, Vec::new()) {
+		for task in std::mem::take(&mut self.background_tasks) {
 			task.join().unwrap();
 		}
 	}
@@ -885,7 +674,7 @@ impl Context {
 				Some(x) => x,
 				_ => return,
 			};
-			if let Err(e) = crate::save_rgba8_image(&path, image.data(), info.width, info.height, info.stride_y) {
+			if let Err(e) = crate::save_rgba8_image(&path, image.data(), info.size, info.stride.y) {
 				log::error!("failed to save image to {}: {}", path, e);
 			}
 		});
@@ -902,7 +691,7 @@ impl Context {
 		let info = image.info();
 		let name = format!("{}.png", name);
 		self.run_background_task(move || {
-			if let Err(e) = crate::save_rgba8_image(&name, image.data(), info.width, info.height, info.stride_y) {
+			if let Err(e) = crate::save_rgba8_image(&name, image.data(), info.size, info.stride.y) {
 				log::error!("failed to save image to {}: {}", name, e);
 			}
 		});
@@ -995,7 +784,7 @@ fn create_window_bind_group_layout(device: &wgpu::Device) -> wgpu::BindGroupLayo
 			ty: wgpu::BindingType::Buffer {
 				ty: wgpu::BufferBindingType::Uniform,
 				has_dynamic_offset: false,
-				min_binding_size: Some(NonZeroU64::new(std::mem::size_of::<WindowUniforms>() as u64).unwrap()),
+				min_binding_size: Some(NonZeroU64::new(WindowUniforms::STD140_SIZE).unwrap()),
 			},
 		}],
 	})
@@ -1042,14 +831,14 @@ fn create_render_pipeline(
 ) -> wgpu::RenderPipeline {
 	device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
 		label: Some("show-image-pipeline"),
-		layout: Some(&layout),
+		layout: Some(layout),
 		vertex: wgpu::VertexState {
-			module: &vertex_shader,
+			module: vertex_shader,
 			entry_point: "main",
 			buffers: &[],
 		},
 		fragment: Some(wgpu::FragmentState {
-			module: &fragment_shader,
+			module: fragment_shader,
 			entry_point: "main",
 			targets: &[wgpu::ColorTargetState {
 				format: swap_chain_format,
@@ -1088,7 +877,7 @@ fn create_render_pipeline(
 
 /// Create a swap chain for a surface.
 fn create_swap_chain(
-	size: winit::dpi::PhysicalSize<u32>,
+	size: glam::UVec2,
 	surface: &wgpu::Surface,
 	format: wgpu::TextureFormat,
 	device: &wgpu::Device,
@@ -1096,12 +885,12 @@ fn create_swap_chain(
 	let swap_chain_desc = wgpu::SwapChainDescriptor {
 		usage: wgpu::TextureUsage::RENDER_ATTACHMENT,
 		format,
-		width: size.width,
-		height: size.height,
+		width: size.x,
+		height: size.y,
 		present_mode: wgpu::PresentMode::Mailbox,
 	};
 
-	device.create_swap_chain(&surface, &swap_chain_desc)
+	device.create_swap_chain(surface, &swap_chain_desc)
 }
 
 /// Perform a render pass of an image.
@@ -1121,7 +910,7 @@ fn render_pass(
 	let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
 		label: Some("render-image"),
 		color_attachments: &[wgpu::RenderPassColorAttachment {
-			view: &target,
+			view: target,
 			resolve_target: None,
 			ops: wgpu::Operations { load, store: true },
 		}],