+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v0.9.0:
+  * Change `WindowHandle` API to provide more efficient access to window properties.
+  * `WindowHandle` now guarantees that the referenced window is valid.
+  * Prohibit access to `&mut ContextHandle` as long as a `WindowHandle` exists.
+  * Rename `show_overlays` property to `overlays_visible`.
+  * Fix interpretation of `tch::Tensor` with planar data.
+  * Add user configurable image transformation.
+  * Add `prev_position` to `MouseMoveEvent`.
+  * Add `position` and `prev_position` to `MouseButtonEvent`.
+  * Add enabled-by-default mouse controls to pan and zoom images.
+  * Use `glam` vectors in public API (and publicly re-export `glam`).
+
 v0.8.5:
   * Update to wgpu `0.9` and winit `0.25`.